requirements.txt
# SuperAI 统一依赖管理
# 整合所有微服务的依赖，解决版本冲突

# === 核心框架 ===
Flask==2.2.2
Werkzeug==2.2.2
gunicorn==23.0.0

# === 数据存储和通信 ===
redis==4.5.5
requests==2.31.0

# === 外部API集成 ===
tavily-python

# === 安全和证书 ===
certifi==2024.7.4

# === 监控和指标 ===
prometheus-client==0.19.0
psutil==5.9.6

# === 开发和调试工具 ===
# 可选：用于开发环境的额外工具
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0

# === 注意事项 ===
# 1. 已移除gevent依赖，统一使用threading模型
# 2. 所有Flask相关包使用相同版本避免冲突
# 3. Redis客户端版本统一为4.5.5
# 4. 生产环境建议固定所有版本号